Issue #30 Resource distribution
Create mechanism of resources distribution. This mechanims allows easy way to provide
resources, when owner of structure changed. The resource network will be not created.

Fix suad fly to direct point of the target (not base of collision) for non-moveable elements.

Append debug functionality of logging information about clicked objects.